refactor(guard): #1218 Step 1c — extract upstream router to app/guard/router.py
Third module of the three-part decomposition. Step 1 complete.

Moves from apps/api/app/modules/guard/routers/proxy.py to
apps/api/app/guard/router.py (verbatim, behavior-identical):
- upstream()          — was _forward (public, provider fanout + streaming + audit)
- _stream_chunks      — private helper (SSE pass-through)
- _safe_json          — private helper (best-effort JSON decode)
- fail_closed()       — was _fail_closed (public, canonical error envelope)

proxy.py keeps compat re-exports under the original names. All internal
call sites of _fail_closed (15+ in proxy.py) work through the re-export.

Regression harness: 8/8 green against docker Postgres.
Downstream: 239 passed.

Progress on #1218 Step 1 — three-module decomposition:
- Step 1a ✓ app/guard/policy.py    — evaluate() + rule matchers
- Step 1b ✓ app/guard/audit.py     — record() + token/cost helpers
- Step 1c ✓ app/guard/router.py    — upstream() + streaming + fail_closed()

Next: Step 2 — app/guard/gateway.py — thin orchestrator composing all three.
