+++ apps/api/app/modules/guard/routers/proxy.py
@@ -81,6 +81,14 @@ from app.guard.audit import (
     record as _record_audit,
 )
 
+# Re-exported from app.guard.router — #1218 Step 1c.
+from app.guard.router import (
+    _safe_json,
+    _stream_chunks,
+    fail_closed as _fail_closed,
+    upstream as _forward,
+)
+
 
 # Action restrictiveness for winner selection when multiple rules match.
 # _ACTION_RANK now re-exported from app.guard.policy (see import block above).
@@ -726,211 +734,6 @@ def _infer_ai_tool(request: Request) -> str:
             return name
     return "unknown"
 
-
-async def _forward(
-    *, upstream: str, path: str, body: dict, real_key: str,
-    auth_header_out: str, bearer: bool, is_stream: bool,
-    background: BackgroundTasks, audit_args: tuple,
-    extra_headers: dict | None = None,
-    upstream_api_key: str | None = None,
-    vendor_key: str | None = None,
-    provider: str = "",
-) -> StreamingResponse | JSONResponse:
-    headers = {
-        "content-type": "application/json",
-        "accept": "text/event-stream" if is_stream else "application/json",
-    }
-
-    # When routing through a BYO gateway, use gateway-specific auth headers.
-    # gateway_adapt returns non-empty headers for known gateways (Portkey, Helicone,
-    # Azure, OpenRouter). Pass the requested model so gateways that need a rewritten
-    # form (OpenRouter, LiteLLM) can produce it; we then patch body["model"] below.
-    _requested_model = body.get("model") or ""
-    if upstream_api_key:
-        from app.runtime.adapters.gateway import gateway_adapt as _gw_adapt
-        _gw = _gw_adapt(upstream, upstream_api_key, provider, _requested_model)
-        if _gw.headers:
-            headers.update(_gw.headers)
-            if vendor_key:
-                headers[auth_header_out] = f"Bearer {vendor_key}" if bearer else vendor_key
-        else:
-            headers[auth_header_out] = f"Bearer {real_key}" if bearer else real_key
-        # Rewrite model in body if the gateway needs a different format
-        # (OpenRouter/LiteLLM prepend "provider/"). Idempotent per adapter.
-        if _gw.model and _requested_model and _gw.model != _requested_model:
-            body["model"] = _gw.model
-    else:
-        headers[auth_header_out] = f"Bearer {real_key}" if bearer else real_key
-
-    # anthropic-version is required for Anthropic-compatible endpoints regardless of upstream
-    if provider == "anthropic":
-        headers.setdefault("anthropic-version", "2023-06-01")
-    if extra_headers:
-        headers.update(extra_headers)
-
-    # Strip whitespace/newlines from all header values — stored keys can have trailing \n
-    headers = {k: str(v).strip() for k, v in headers.items()}
-
-    # Build full URL: strip any path prefix the upstream already contains so
-    # upstream="/v1" + path="/v1/messages" doesn't produce double /v1/v1/messages.
-    _up_path = _urlparse(upstream).path.rstrip("/")
-    _req_path = path[len(_up_path):] if _up_path and path.startswith(_up_path) else path
-    _full_url = upstream.rstrip("/") + _req_path
-
-    # ponytail: a single shared async client would be better for connection
-    # pooling. Per-call is fine until QPS warrants it.
-    _breaker = _get_breaker()
-    _breaker_key = provider or "default"
-    if not _breaker.allow(_breaker_key):
-        log.warning("guard.proxy.breaker_open", provider=_breaker_key,
-                    snapshot=_breaker.snapshot(_breaker_key))
-        return _fail_closed(
-            503,
-            f"Guard circuit breaker OPEN for provider={_breaker_key} — "
-            f"upstream failing repeatedly; retry after ~{int(_breaker.recovery_timeout)}s",
-        )
-
-    client = httpx.AsyncClient(timeout=httpx.Timeout(600.0))
-
-    try:
-        req = client.build_request("POST", _full_url, json=body, headers=headers)
-        log.info("guard.proxy.forward", url=_full_url,
-                 headers={k: v for k, v in headers.items() if "key" not in k.lower() and "auth" not in k.lower()})
-        resp = await client.send(req, stream=True)
-    except Exception as e:
-        _breaker.record_failure(_breaker_key)
-        await client.aclose()
-        import traceback as _tb
-        log.warning("guard.proxy.upstream_unreachable", url=_full_url,
-                    exc_type=type(e).__name__, err=str(e),
-                    traceback=_tb.format_exc())
-        return _fail_closed(502, f"Upstream {_full_url} unreachable: {type(e).__name__}: {e}")
-
-    # 5xx = upstream failure counted against the breaker; 4xx = client error, ignored.
-    if resp.status_code >= 500:
-        _breaker.record_failure(_breaker_key)
-    elif resp.status_code < 400:
-        _breaker.record_success(_breaker_key)
-
-    if resp.status_code >= 400:
-        # Read the error body, close client, return as-is so the SDK sees the
-        # real vendor error.
-        err_body = await resp.aread()
-        await resp.aclose()
-        await client.aclose()
-        return JSONResponse(
-            status_code=resp.status_code,
-            content=_safe_json(err_body, fallback={"error": "upstream error"}),
-        )
-
-    if is_stream:
-        return StreamingResponse(
-            _stream_chunks(client, resp, background, audit_args, upstream=upstream),
-            media_type=resp.headers.get("content-type", "text/event-stream"),
-        )
-
-    # Non-streaming path
-    full = await resp.aread()
-    # httpx with stream=True returns raw bytes; content-encoding is NOT
-    # auto-decoded. Anthropic responds with gzip/brotli when we forward the
-    # SDK's accept-encoding header. Decompress here or _safe_json downstream
-    # sees binary garbage, falls back to {}, and every brain block returns
-    # an empty envelope silently.
-    _enc = (resp.headers.get("content-encoding") or "").lower().strip()
-    if _enc == "gzip":
-        try:
-            import gzip as _gzip
-            full = _gzip.decompress(full)
-        except Exception as e:
-            log.warning("guard.proxy.gzip_decompress_failed", err=str(e))
-    elif _enc in ("br", "brotli"):
-        try:
-            import brotli as _brotli
-            full = _brotli.decompress(full)
-        except ImportError:
-            log.warning("guard.proxy.brotli_missing", note="pip install brotli")
-        except Exception as e:
-            log.warning("guard.proxy.brotli_decompress_failed", err=str(e))
-    elif _enc == "deflate":
-        try:
-            import zlib as _zlib
-            full = _zlib.decompress(full)
-        except Exception as e:
-            log.warning("guard.proxy.deflate_decompress_failed", err=str(e))
-    await resp.aclose()
-    await client.aclose()
-    background.add_task(
-        _record_audit, *audit_args[:6], audit_args[6],
-        int((time.monotonic() - audit_args[7]) * 1000),
-        body=audit_args[8], response_bytes=full, upstream=upstream,
-        prompt_summary=audit_args[9] if len(audit_args) > 9 else "",
-        user_email=audit_args[10] if len(audit_args) > 10 else None,
-        conductai_run_id=audit_args[11] if len(audit_args) > 11 else None,
-        conductai_workflow=audit_args[12] if len(audit_args) > 12 else None,
-        conductai_workflow_id=audit_args[13] if len(audit_args) > 13 else None,
-        hook_session_id=audit_args[14] if len(audit_args) > 14 else None,
-    )
-    # Drop the now-stale content-encoding header so downstream (the SDK
-    # inside brain_block) doesn't try to decompress the already-decompressed
-    # body a second time.
-    _resp_headers = {k: v for k, v in resp.headers.items() if k.lower() != "content-encoding" and k.lower() != "content-length"}
-    return JSONResponse(
-        status_code=resp.status_code,
-        content=_safe_json(full, fallback={}),
-        headers=_resp_headers,
-    )
-
-
-async def _stream_chunks(
-    client: httpx.AsyncClient, resp: httpx.Response,
-    background: BackgroundTasks, audit_args: tuple,
-    upstream: str | None = None,
-) -> AsyncIterator[bytes]:
-    """Pass-through every chunk. Schedule the audit event after the stream
-    closes — we don't parse mid-stream in V1."""
-    collected = bytearray()
-    try:
-        async for chunk in resp.aiter_bytes():
-            collected.extend(chunk)
-            yield chunk
-    finally:
-        await resp.aclose()
-        await client.aclose()
-        background.add_task(
-            _record_audit, *audit_args[:6], audit_args[6],
-            int((time.monotonic() - audit_args[7]) * 1000),
-            body=audit_args[8], response_bytes=bytes(collected), upstream=upstream,
-            prompt_summary=audit_args[9] if len(audit_args) > 9 else "",
-            user_email=audit_args[10] if len(audit_args) > 10 else None,
-            conductai_run_id=audit_args[11] if len(audit_args) > 11 else None,
-            conductai_workflow=audit_args[12] if len(audit_args) > 12 else None,
-            conductai_workflow_id=audit_args[13] if len(audit_args) > 13 else None,
-            hook_session_id=audit_args[14] if len(audit_args) > 14 else None,
-        )
-
-
-def _safe_json(b: bytes, *, fallback: dict) -> dict:
-    try:
-        return json.loads(b)
-    except Exception:
-        # Log a snippet so operators can see WHY the upstream returned junk —
-        # silent fallback to {} looks like a valid empty response to the SDK
-        # and turns into an all-hands debug session later.
-        try:
-            snippet = (b[:400] or b"").decode("utf-8", errors="replace")
-        except Exception:
-            snippet = repr(b[:400])
-        log.warning("guard.proxy.upstream_unparseable_body",
-                    snippet=snippet, byte_length=len(b or b""))
-        return fallback
-
-
-def _fail_closed(status: int, message: str) -> JSONResponse:
-    """Security tool failing open is worse than no tool. Surface clear errors."""
-    return JSONResponse(
-        status_code=status,
-        content={"error": {"type": "conduct_guard_proxy", "message": message}},
-    )
 
 
 # ── Proxy config endpoints ────────────────────────────────────────────────────
